DSND-2170: Move service result into service package
