+++ docs/TM01_put_request_body.json
@@ -3,23 +3,23 @@
     "links": {
       "self": "/company/12345678/filing-history/MzA1MzY0NzkxNWFkaXF6a2N4"
     },
-    "category" : "officers",
-    "subcategory" : "termination",
-    "action_date" : "2011-08-06",
-    "date" : "2011-09-05T05:39:19.000Z",
-    "type" : "TM01",
-    "description_values" : {
-      "officer_name" : "John Tester",
-      "termination_date" : "2011-08-06"
+    "category": "officers",
+    "subcategory": "termination",
+    "action_date": "2011-08-06",
+    "date": "2011-09-05T05:39:19.000Z",
+    "type": "TM01",
+    "description_values": {
+      "officer_name": "John Tester",
+      "termination_date": "2011-08-06"
     },
-    "description" : "termination-director-company-with-name-termination-date"
+    "description": "termination-director-company-with-name-termination-date"
   },
   "internal_data": {
     "original_values": {
-      "resignation_date" : "06/08/2011",
+      "resignation_date": "06/08/2011",
       "officer_name": "John Tester"
     },
-    "original_description":"Appointment terminated, director john tester",
+    "original_description": "Appointment terminated, director john tester",
     "company_number": "12345678",
     "entity_id": "3053647915",
     "document_id": "000X4BI89B65846",
