+++ src/main/java/uk/gov/companieshouse/filinghistory/api/controller/FilingHistoryController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
 import uk.gov.companieshouse.filinghistory.api.logging.DataMapHolder;
-import uk.gov.companieshouse.filinghistory.api.model.ServiceResult;
+import uk.gov.companieshouse.filinghistory.api.service.ServiceResult;
 import uk.gov.companieshouse.filinghistory.api.service.Processor;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
