+++ src/main/java/uk/gov/companieshouse/filinghistory/api/service/Processor.java
@@ -1,7 +1,6 @@
 package uk.gov.companieshouse.filinghistory.api.service;
 
 import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
-import uk.gov.companieshouse.filinghistory.api.model.ServiceResult;
 
 public interface Processor {
