+++ src/test/java/uk/gov/companieshouse/filinghistory/api/service/FilingHistoryProcessorTest.java
@@ -19,7 +19,6 @@ import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
 import uk.gov.companieshouse.filinghistory.api.mapper.AbstractTransactionMapperFactory;
 import uk.gov.companieshouse.filinghistory.api.mapper.TopLevelTransactionMapper;
 import uk.gov.companieshouse.filinghistory.api.model.FilingHistoryDocument;
-import uk.gov.companieshouse.filinghistory.api.model.ServiceResult;
 
 @ExtendWith(MockitoExtension.class)
 class FilingHistoryProcessorTest {
